+++ src/Graphify/Semantics/IPropertySymbolExtensions.ToProperty.cs
@@ -13,15 +13,15 @@
         /// </summary>
         /// <param name="property">The property symbol to convert. Cannot be <see langword="null"/>.</param>
         /// <returns>A <see cref="Property"/> instance representing the specified property symbol.</returns>
-        public static Property ToProperty(this IPropertySymbol property)
+        public static Property ToProperty(this IPropertySymbol property, TraverseScope scope)
         {
             _ = property.Type.IsSequence(out Element element);
 
             return new Property
             {
                 Element = element,
-                IsIgnored = false,
                 Name = property.Name,
+                Scope = scope,
                 Symbol = property.Type,
                 Type = property.Type.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat),
             };