+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [1.0.0] - TBC
 
 - Initial Release
+- Renamed the Ignore attribute to Traverse and added TraverseScope to control traversal behavior.